+++ CryptoCheck.Core/Models/ExchangeRates.cs
@@ -8,10 +8,13 @@ namespace CryptoCheck.Core.Models
     {
         public long Timestamp { get; set; }
 
+        /// <summary>
+        /// The symbol of the currency used as the base currency (eg. EUR)
+        /// </summary>
         public string BaseSymbol { get; set; }
 
         public DateTimeOffset Date { get; set; }
 
-        public Dictionary<string, double> Rates { get; set; }
+        public Dictionary<string, decimal> Rates { get; set; }
     }
 }
