+++ CryptoCheck.Services/ExchangeRatesApi/Models/ExchangeRates.cs
@@ -20,6 +20,6 @@ namespace CryptoCheck.Services.ExchangeRatesApi.Models
         public DateTimeOffset Date { get; set; }
 
         [JsonProperty("rates")]
-        public Dictionary<string, double> Rates { get; set; }
+        public Dictionary<string, decimal> Rates { get; set; }
     }
 }
